add and configure assets and images; adjust Appointment flow
