+++ lib/domain/entities/visit_history_model.dart
