+++ lib/infra/helpers/consultation_type.dart
