+++ lib/presentation/components/doctor_list.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:smart_med/infra/mock/mock_doctors.dart';
 import 'package:smart_med/presentation/doctor_card/doctor_card.dart';
-import 'package:smart_med/presentation/views/doctor_details_screen.dart';
+import 'package:smart_med/presentation/views/create_appointment_screen.dart';
 
 Widget DoctorList(BuildContext context) {
   return Column(
@@ -12,11 +12,9 @@ Widget DoctorList(BuildContext context) {
             child: DoctorCard(
               doctor: doctor,
               onTap: () {
-                Navigator.of(context).push(
-                  MaterialPageRoute(
-                    builder: (_) => DoctorDetailsScreen(doctorId: doctor.id),
-                  ),
-                );
+                Navigator.of(
+                  context,
+                ).push(MaterialPageRoute(builder: (_) => CreateAppointmentScreen()));
               },
             ),
           ),
