+++ lib/presentation/navigation/shell_screen.dart
@@ -4,9 +4,8 @@ import 'package:smart_med/presentation/navigation/app_router.dart';
 import 'package:smart_med/presentation/navigation/smart_med_navigation_bar.dart';
 import 'package:smart_med/presentation/views/account/account_screen.dart';
 import 'package:smart_med/presentation/views/home_screen.dart';
-import 'package:smart_med/presentation/views/appointment_screen.dart' hide AccountScreen;
+import 'package:smart_med/presentation/views/appointment_screen.dart';
 import 'package:smart_med/presentation/views/history_screen.dart';
-import 'package:smart_med/presentation/views/personal_appointment_screen.dart';
 
 class ShellScreen extends StatelessWidget {
   final BeamState routerState;
@@ -25,8 +24,8 @@ class ShellScreen extends StatelessWidget {
         physics: const NeverScrollableScrollPhysics(),
         children: [
           const HomeScreen(),
-          PersonalAppointmentScreen(),
-          //AppointmentScreen(),
+          //PersonalAppointmentScreen(),
+          AppointmentScreen(),
           const HistoryScreen(),
           AccountScreen(),
         ],
